Create dedicated file for opaque type tests
